+++ tests/testthat/test-union.R
@@ -35,8 +35,7 @@ test_that("base unions display as expected", {
 })
 
 test_that("can construct from S3 and S4 classes", {
-  S4_union <- methods::setClass("S4_union")
-  on.exit(S4_remove_classes("S4_union"))
+  S4_union := local_S4_class()
 
   u <- new_union(class_factor, S4_union)
   expect_equal(u$classes, list(class_factor, getClass("S4_union")))
@@ -44,10 +43,9 @@ test_that("can construct from S3 and S4 classes", {
 
 test_that("can construct with |", {
   foo := new_class()
-  Foo1 <- setClass("Foo1", slots = list("x" = "numeric"))
-  Foo2 <- setClass("Foo2", slots = list("x" = "numeric"))
-  Foo3 <- setClassUnion("Foo3", c("Foo1", "Foo2"))
-  on.exit(S4_remove_classes(c("Foo1", "Foo2", "Foo3")))
+  Foo1 := local_S4_class(slots = list("x" = "numeric"))
+  Foo2 := local_S4_class(slots = list("x" = "numeric"))
+  Foo3 := local_S4_union(c("Foo1", "Foo2"))
 
   expect_equal(class_integer | class_double, class_numeric)
   expect_equal(class_integer | class_numeric, class_numeric)